requirements.txt
fastmcp>=0.1.0
anthropic>=0.25.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
